+++ main.go
@@ -103,7 +103,6 @@ func main() {
 	rootCmd.PersistentFlags().BoolP("add-global-variants", "", false, "Enable generation for global variants")
 	rootCmd.PersistentFlags().BoolP("add-panic-variants", "", false, "Enable generation for panic variants")
 	rootCmd.PersistentFlags().BoolP("add-soft-deletes", "", false, "Enable soft deletion by updating deleted_at timestamp")
-	rootCmd.PersistentFlags().BoolP("add-enum-types", "", false, "Enable generation of types for enums")
 	rootCmd.PersistentFlags().StringP("enum-null-prefix", "", "Null", "Name prefix of nullable enum types")
 	rootCmd.PersistentFlags().BoolP("version", "", false, "Print the version")
 	rootCmd.PersistentFlags().BoolP("wipe", "", false, "Delete the output folder (rm -rf) before generation to ensure sanity")
@@ -158,7 +157,6 @@ func preRun(cmd *cobra.Command, args []string) error {
 		AddGlobal:         viper.GetBool("add-global-variants"),
 		AddPanic:          viper.GetBool("add-panic-variants"),
 		AddSoftDeletes:    viper.GetBool("add-soft-deletes"),
-		AddEnumTypes:      viper.GetBool("add-enum-types"),
 		EnumNullPrefix:    viper.GetString("enum-null-prefix"),
 		NoTests:           viper.GetBool("no-tests"),
 		NoHooks:           viper.GetBool("no-hooks"),
@@ -200,7 +198,6 @@ func preRun(cmd *cobra.Command, args []string) error {
 	cmdConfig.DriverConfig = map[string]interface{}{
 		"whitelist":        viper.GetStringSlice(driverName + ".whitelist"),
 		"blacklist":        viper.GetStringSlice(driverName + ".blacklist"),
-		"add-enum-types":   cmdConfig.AddEnumTypes,
 		"enum-null-prefix": cmdConfig.EnumNullPrefix,
 	}